Comply with new variant system
To comply with the new variant system there needs to be a feature for
acrn and the multiconfig needs to be moved.

The choice of putting multiconfig in conf/multiconfig_sample is somewhat
arbitray. I do not feel to should be in conf/sample and
multiconfig_sample feels better. The word sample is used as it should be
copied to build/conf in the same way as the contents of conf/sample
should.